requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
cached-property==1.5.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
configparser==5.0.2
cssselect==1.1.0
discord.py==1.7.1
discord==1.0.1
fake-useragent==0.1.11
feedparser==6.0.2
idna==2.10
importlib-metadata==2.1.1; python_version < '3.8'
lxml==4.6.3
multidict==5.1.0
multitasking==0.0.9
numpy==1.20.2
pandas==1.2.3
parse==1.19.0
parsedatetime==2.6
pathlib==1.0.1
poker==0.30.0
pyee==8.1.0
pyppeteer==0.2.5
pyquery==1.4.3
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
requests-html==0.10.0
requests==2.25.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2.1; python_version >= '3.0'
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
w3lib==1.22.0
websockets==8.1
yahoo-fin==0.8.8
yahoofinancials==1.6
yarl==1.6.3
yfinance==0.1.59
zipp==3.4.1
zope.interface==5.3.0
